agent/requirements.txt
# Second Brain Agent - Python Dependencies
# Use `uv sync` (recommended) or `pip install -r requirements.txt` (fallback)

pydantic-ai>=0.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
ag-ui-protocol>=0.1.0
httpx>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
ruff>=0.6.0
